requirements.txt
django=latest
pillow=latest
sorl-thumbnail=latest